Provide fallback for IntlProvider (#361)
Currently the application cannot serve a correct 404. When calling an
URL with a wrong language identifier (e.g. /foo instead /en) the page is
being rendered and thus the layouts as well. However, the layout returns
a 500 because the messages can't be loaded.

With this change the 404 is returned in english if the language can't be
resolved.