Prefer mapped types over column types when possible (#376)
## Description

I've noticed that we're currently doing this inconsistently while
debugging a bug in
[getCrudSearchFieldsFromMetadata](https://github.com/vivid-planet/comet/blob/main/packages/api/cms-api/src/generator/utils/search-fields-from-metadata.ts#L6)
together with @andrearutrecht. Since we're checking for `type` there,
and not `columnType`, we should prefer `type` over `columnType` (We
still need to fix the bug though).

## Further information

I've also looked at the MikroORM docs to find a best practice and found
that they're using `type` in the
[tutorial](https://mikro-orm.io/docs/guide/first-entity#first-entity),
so I believe using `type` is preferred when it's possible.