Add RootLayout (#407)
Currently, there is a layout file for the site and a layout file for the block preview in the project. These partially contain the same content, for example both of them define the html and body tags.

To avoid code duplication, a root layout is created at the top-most level of the app directory. The existing layouts are used as child layouts of this one.

This way, fonts and similar resources no longer need to be imported twice.

Co-authored-by: Stefanie Kaltenhauser <stefanie.kaltenhauser@vivid-planet.com>