Site: fix broken NODE_OPTIONS for dev script
Demo PR: https://github.com/vivid-planet/comet/pull/4265

## Old

NODE_OPTIONS where passed to dotenv (which is a node script itself), when attaching the debugger to 9230 you connected to the dotenv script, not the site script. The --max-old-space-size also had no effect.

## Now

As we start node ourself (without wrapper) we simply give the options directly to node.